requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
python-dotenv>=1.0
tqdm>=4.66
beautifulsoup4>=4.12
lxml>=5.0
rapidfuzz>=3.6
tenacity>=8.2
